backend/ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.0.1
pandas==2.1.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.350
torch==2.1.1
huggingface-hub==0.19.4
python-jose==3.3.0
pydantic==2.5.2 